utils/HelperFunctions.hpp
